+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **doc:** fix flaw
 - **fix:** exposing of result
 - **add:** tests for validate.options
+- **add:** tests for validate.query
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure